requirements.txt

gymnasium
numpy
